# Conflicts:
#	data/lib/repository/card/card_repository_impl.dart
#	data/lib/source/card/card_datasource.dart
#	domain/lib/repository/card/card_repository.dart
#	lib/di/usecase/card_delivery/card_delivery_usecase_provider.dart
#	lib/feature/register/stepone/enter_address/enter_address_view.dart
#	lib/feature/set_credit_limit/set_credit_limit_page_view.dart
#	lib/generated/intl/messages_en.dart
#	lib/ui/molecules/dialog/register/step_four/state_city_dialog/state_city_dialog_view.dart
#	pubspec.yaml

+++ lib/feature/payment/request_amount_from_contact/request_amount_from_contact_view_model.dart
@@ -164,7 +164,7 @@ class RequestAmountFromContactViewModel extends BasePageViewModel {
               ? 'Transfer to Friend or Family'
               : beneficiary!.purposeDetails!)
           : purposeDetail!.strCode!,
-      amount: double.parse(currentPinValue).toInt(),
+      amount: double.parse(currentPinValue),
       dbtrBic: beneficiary!.iban ?? "",
       dbtrAcct: beneficiary!.accountNo ?? "",
       dbtrName: beneficiary!.fullName ?? "",
